[ERM-4706] При попытке очистить пустой список в [PerformedOperationFinalProcessings] возникает WARN
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Prerelease/Platform;C23345

+++ Source/ApplicationServices/2Gis.Erm.Platform.Aggregates/SimplifiedModel/PerformedOperations/Operations/OperationsFinalProcessingCompleteAggregateService.cs
@@ -1,8 +1,10 @@
 ﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Transactions;
 
 using DoubleGis.Erm.Platform.API.Aggregates.SimplifiedModel.PerformedOperations.Operations;
+using DoubleGis.Erm.Platform.Common.Utils.Data;
 using DoubleGis.Erm.Platform.DAL.PersistenceServices;
 using DoubleGis.Erm.Platform.DAL.Transactions;
 using DoubleGis.Erm.Platform.Model.Entities.Erm;
@@ -18,22 +20,28 @@ namespace DoubleGis.Erm.Platform.Aggregates.SimplifiedModel.PerformedOperations.
             _batchDeletePersistenceService = batchDeletePersistenceService;
         }
 
-        public void CompleteProcessing(IEnumerable<PerformedOperationFinalProcessing> completedProcessing)
+        public void CompleteProcessing(IReadOnlyList<PerformedOperationFinalProcessing> completedProcessing)
         {
+            if (completedProcessing.Count == 0)
+            {
+                return;
+            }
+
             const int BatchSize = 10000;
-            int offset = 0;
 
             using (var transaction = new TransactionScope(TransactionScopeOption.Required, DefaultTransactionOptions.Default))
             {
-                IEnumerable<PerformedOperationFinalProcessing> batch;
-                do
+                for (int offset = 0; offset < completedProcessing.Count;)
                 {
-                    batch = completedProcessing.Skip(offset).Take(BatchSize);
-                    offset += BatchSize;
+                    var batch = completedProcessing.SkipTake(offset, BatchSize);
+                    if (batch.Count == 0)
+                    {
+                        break;
+                    }
 
                     _batchDeletePersistenceService.Delete(batch);
+                    offset += batch.Count;
                 }
-                while (batch.Any());
 
                 transaction.Complete();
             }
